dspy/requirements.txt
dspy
fastapi
uvicorn
httpx
pandas
openai
python-dotenv 